Refatora serviços de execução por etapa no Gera Landing
