+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/copy/CopyExecutionScheduler.java
@@ -1,6 +1,5 @@
 package com.marketinghub.worker.geralanding.copy;
 
-import com.marketinghub.worker.geralanding.comum.GeraLandingCopyExecutionService;
 
 
 import org.slf4j.Logger;
