+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/copy/GeraLandingCopyExecutionService.java
@@ -1,20 +1,20 @@
-package com.marketinghub.worker.geralanding.comum;
+package com.marketinghub.worker.geralanding.copy;
 
 import com.marketinghub.worker.geralanding.GeraLandingExecutionService;
 import com.marketinghub.worker.geralanding.GeraLandingStageExecutionDto;
 import java.util.List;
 import org.springframework.stereotype.Service;
 
-/** Centraliza a execução de jobs da etapa preset design usando o executor compartilhado. */
+/** Centraliza a execução de jobs da etapa copy usando o executor compartilhado. */
 @Service
-public class GeraLandingPresetDesignExecutionService {
+public class GeraLandingCopyExecutionService {
     private final GeraLandingExecutionService executionService;
 
-    public GeraLandingPresetDesignExecutionService(GeraLandingExecutionService executionService) {
+    public GeraLandingCopyExecutionService(GeraLandingExecutionService executionService) {
         this.executionService = executionService;
     }
 
-    /** Processa os jobs pendentes da etapa preset design. */
+    /** Processa os jobs pendentes da etapa copy. */
     public void processExecutions(List<GeraLandingStageExecutionDto> jobs) {
         executionService.processExecutions(jobs);
     }
