+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/deliverables/DeliverablesExecutionScheduler.java
@@ -1,6 +1,5 @@
 package com.marketinghub.worker.geralanding.deliverables;
 
-import com.marketinghub.worker.geralanding.comum.GeraLandingDeliverablesExecutionService;
 
 
 import org.slf4j.Logger;
