+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/imageplanning/ImagePlanningExecutionScheduler.java
@@ -1,6 +1,5 @@
 package com.marketinghub.worker.geralanding.imageplanning;
 
-import com.marketinghub.worker.geralanding.comum.GeraLandingImagePlanningExecutionService;
 
 
 import org.slf4j.Logger;
