+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/presetdesign/GeraLandingPresetDesignExecutionService.java
@@ -1,20 +1,20 @@
-package com.marketinghub.worker.geralanding.comum;
+package com.marketinghub.worker.geralanding.presetdesign;
 
 import com.marketinghub.worker.geralanding.GeraLandingExecutionService;
 import com.marketinghub.worker.geralanding.GeraLandingStageExecutionDto;
 import java.util.List;
 import org.springframework.stereotype.Service;
 
-/** Centraliza a execução de jobs da etapa deliverables usando o executor compartilhado. */
+/** Centraliza a execução de jobs da etapa preset design usando o executor compartilhado. */
 @Service
-public class GeraLandingDeliverablesExecutionService {
+public class GeraLandingPresetDesignExecutionService {
     private final GeraLandingExecutionService executionService;
 
-    public GeraLandingDeliverablesExecutionService(GeraLandingExecutionService executionService) {
+    public GeraLandingPresetDesignExecutionService(GeraLandingExecutionService executionService) {
         this.executionService = executionService;
     }
 
-    /** Processa os jobs pendentes da etapa deliverables. */
+    /** Processa os jobs pendentes da etapa preset design. */
     public void processExecutions(List<GeraLandingStageExecutionDto> jobs) {
         executionService.processExecutions(jobs);
     }
