+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/presetdesign/PresetDesignExecutionScheduler.java
@@ -1,6 +1,5 @@
 package com.marketinghub.worker.geralanding.presetdesign;
 
-import com.marketinghub.worker.geralanding.comum.GeraLandingPresetDesignExecutionService;
 
 
 import org.slf4j.Logger;
