+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/wireframe/GeraLandingWireframeExecutionService.java
@@ -1,20 +1,20 @@
-package com.marketinghub.worker.geralanding.comum;
+package com.marketinghub.worker.geralanding.wireframe;
 
 import com.marketinghub.worker.geralanding.GeraLandingExecutionService;
 import com.marketinghub.worker.geralanding.GeraLandingStageExecutionDto;
 import java.util.List;
 import org.springframework.stereotype.Service;
 
-/** Centraliza a execução de jobs da etapa copy usando o executor compartilhado. */
+/** Centraliza a execução de jobs da etapa wireframe usando o executor compartilhado. */
 @Service
-public class GeraLandingCopyExecutionService {
+public class GeraLandingWireframeExecutionService {
     private final GeraLandingExecutionService executionService;
 
-    public GeraLandingCopyExecutionService(GeraLandingExecutionService executionService) {
+    public GeraLandingWireframeExecutionService(GeraLandingExecutionService executionService) {
         this.executionService = executionService;
     }
 
-    /** Processa os jobs pendentes da etapa copy. */
+    /** Processa os jobs pendentes da etapa wireframe. */
     public void processExecutions(List<GeraLandingStageExecutionDto> jobs) {
         executionService.processExecutions(jobs);
     }
