+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/wireframe/WireframeExecutionScheduler.java
@@ -1,6 +1,5 @@
 package com.marketinghub.worker.geralanding.wireframe;
 
-import com.marketinghub.worker.geralanding.comum.GeraLandingWireframeExecutionService;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
